feat(config): get kadena directory config service + fix circular dep
