+++ packages/tools/kadena-cli/src/account/accountOptions.ts
@@ -5,7 +5,7 @@ import { CHAIN_ID_RANGE_ERROR_MESSAGE } from '../constants/account.js';
 import { actionAskForDeployFaucet } from '../prompts/genericActionPrompts.js';
 import { account } from '../prompts/index.js';
 import { createOption } from '../utils/createOption.js';
-import { formatZodError, generateAllChainIds } from '../utils/helpers.js';
+import { formatZodError, generateAllChainIds } from '../utils/globalHelpers.js';
 import { log } from '../utils/logger.js';
 import type { IAliasAccountData } from './types.js';
 import {
