+++ packages/tools/kadena-cli/src/account/commands/accountDelete.ts
@@ -5,7 +5,7 @@ import { KadenaError } from '../../services/service-error.js';
 import type { CommandResult } from '../../utils/command.util.js';
 import { assertCommandError } from '../../utils/command.util.js';
 import { createCommand } from '../../utils/createCommand.js';
-import { isNotEmptyString } from '../../utils/helpers.js';
+import { isNotEmptyString } from '../../utils/globalHelpers.js';
 import { log } from '../../utils/logger.js';
 import { accountOptions } from '../accountOptions.js';
 import {
