+++ packages/tools/kadena-cli/src/prompts/wallets.ts
@@ -2,7 +2,7 @@ import { services } from '../services/index.js';
 import type { IWallet } from '../services/wallet/wallet.types.js';
 import { CommandError } from '../utils/command.util.js';
 
-import { isValidFilename } from '../utils/helpers.js';
+import { isValidFilename } from '../utils/globalHelpers.js';
 import { input, select } from '../utils/prompts.js';
 
 export async function walletNamePrompt(): Promise<string> {
